requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI APIs (Whisper + GPT)
openai==1.3.7

# Data Processing & Visualization
plotly==5.18.0
pandas==2.1.3

# Templating
jinja2==3.1.2
aiofiles==23.2.1

# Environment & Validation
python-dotenv==1.0.0
pydantic==2.5.2

# Database (SQLAlchemy + PostgreSQL)
sqlalchemy==2.0.23
aiosqlite==0.19.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HTTP Client (version constrained by supabase)
httpx==0.24.1

# Supabase Storage
supabase==1.2.0
storage3==0.5.3

# Vercel Serverless Adapter
mangum==0.17.0
